requirements.txt
fastapi
uvicorn
motor
pymongo
python-jose[cryptography]
bcrypt
python-dotenv
